Store time as seconds as a long in map store
This avoids overflowing the value in 2038.

Closes #10960

+++ model/map-hot-rod/src/main/java/org/keycloak/models/map/storage/hotRod/authSession/HotRodRootAuthenticationSessionEntity.java
@@ -49,7 +49,7 @@ public class HotRodRootAuthenticationSessionEntity extends AbstractHotRodEntity 
     public String realmId;
 
     @ProtoField(number = 4)
-    public Integer timestamp;
+    public Long timestamp;
 
     @ProtoDoc("@Field(index = Index.YES, store = Store.YES)")
     @ProtoField(number = 5)
